Update Command backend

+++ cu-manager/src/main/java/fr/treeptik/cloudunit/service/impl/CommandServiceImpl.java
@@ -2,15 +2,15 @@ package fr.treeptik.cloudunit.service.impl;
 
 import fr.treeptik.cloudunit.dao.CommandDAO;
 import fr.treeptik.cloudunit.dto.CommandRequest;
+import fr.treeptik.cloudunit.exception.ServiceException;
 import fr.treeptik.cloudunit.model.Command;
 import fr.treeptik.cloudunit.service.CommandService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.inject.Inject;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 public class CommandServiceImpl implements CommandService {
@@ -20,40 +20,62 @@ public class CommandServiceImpl implements CommandService {
 
     @Override
     @Transactional
-    public void addCommand(CommandRequest commandRequest) {
-        Set<String> arguments = new HashSet<String>(commandRequest.getArguments());
+    public void addCommand(CommandRequest commandRequest) throws ServiceException {
+        if(commandRequest.getValue() == null)
+            throw new ServiceException("The value is empty");
+
+        if(commandRequest.getDescription() == null)
+            throw new ServiceException("The description is empty");
+
         Command command = new Command();
+
         command.setValue(commandRequest.getValue());
-        command.setArguments(arguments);
+        command.setArguments(commandRequest.getArguments());
         command.setContainerId(commandRequest.getContainerId());
         commandDAO.save(command);
     }
 
     @Override
     @Transactional
-    public void deleteCommand(Integer id) {
+    public void deleteCommand(Integer id) throws ServiceException {
+        if(id == null)
+            throw new ServiceException("Id is null");
+
         commandDAO.delete(id);
     }
 
     @Override
     @Transactional
-    public void updateCommand(CommandRequest commandRequest) {
-        Set<String> arguments = new HashSet<String>(commandRequest.getArguments());
+    public void updateCommand(CommandRequest commandRequest) throws ServiceException {
+        if(commandRequest.getValue() == null)
+            throw new ServiceException("The value is empty");
+
+        if(commandRequest.getDescription() == null)
+            throw new ServiceException("The description is empty");
+
         Command command = new Command();
         command.setId(commandRequest.getId());
         command.setValue(commandRequest.getValue());
-        command.setArguments(arguments);
+        command.setArguments(commandRequest.getArguments());
         command.setContainerId(commandRequest.getContainerId());
         commandDAO.save(command);
     }
 
     @Override
-    public List<Command> listCommandByContainer(String containerId) {
-        return commandDAO.findByContainer(containerId);
+    public List<CommandRequest> listCommandByContainer(String containerId) throws ServiceException {
+        if(containerId == null)
+            throw new ServiceException("The container id is empty");
+
+        List<Command> commandList = commandDAO.findByContainer(containerId);
+
+        return commandList.stream().map(v -> v.mapToRequest()).collect(Collectors.toList());
     }
 
     @Override
-    public Command getCommand(Integer id) {
-        return commandDAO.findById(id);
+    public CommandRequest getCommand(Integer id) throws ServiceException {
+        if(id == null)
+            throw new ServiceException("The id is empty");
+
+        return commandDAO.findById(id).mapToRequest();
     }
 }
